modifique todo los templates
