+++ src/assets/js/validation.js
@@ -1,9 +1,11 @@
 import{createAccount} from './../js/auth.js'
+import{writeUserData} from './../js/auth.js'
 export const validate = (userName,userLastName,email,password) =>{
     if (userName ===""||userLastName===""||email===""||password===""){
         alert('debe completar los datos');
     }
     else {
         createAccount(email,password);
+        writeUserData(userName,userLastName,email)
     }
 }
