--- src/assets/views/templateLogOut.js
@@ -1,19 +1,0 @@
-import{logOut} from './../js/auth.js';
-
-export const templateLogOut =() => {
-  
-    const containerLogOut = document.createElement('div');
-
-    const contentlogOut = `<p>Desloguearse</p>
-                                <button id="LogOut">Desloguearse</button>`
-
-    containerLogOut.innerHTML = contentlogOut;
-    const btn=containerLogOut.querySelector('#LogOut');
-
-    btn.addEventListener('click',()=>{
-        logOut();
-            
-})
-
-return containerLogOut;
-}
