+++ src/main.js
@@ -11,6 +11,9 @@ var config = {
   };
   firebase.initializeApp(config);
 
+ //Get a reference to the database service
+ var database = firebase.database();
+
 import {initRouter} from './route.js';
 
 const init =()=>{
